extract `lib.user.edit`
